streamlined TS files, adjusted CSS for pickr
