+++ {{cookiecutter.repo_name}}/{{cookiecutter.repo_name}}/static_source/css/base/_forms.scss
@@ -103,6 +103,7 @@ $background-color: #fff;
     & ~ label {
       @apply px-4;
     }
+
     & input:focus {
       @apply ring-0;
     }
@@ -220,11 +221,17 @@ $background-color: #fff;
   select:-webkit-autofill,
   select:-webkit-autofill:hover,
   select:-webkit-autofill:focus {
+    @apply outline-none shadow-none;
     background-clip: text;
     font-size: inherit;
+    box-shadow: none;
   }
 }
 
+// styling for flatpickr (generally fine except overriding focus styles)
+.flatpickr-input {
+  @apply focus:ring-0;
+}
 // global/admin styling for tom-select
 .ts-wrapper {
   @apply px-4 ring-0;
