+++ {{cookiecutter.repo_name}}/{{cookiecutter.repo_name}}/templates/forms/date_datetime.jinja
@@ -3,7 +3,7 @@
     {% set event_name = "update-"|join(id | replace('_', '-')) if id else "" %}
     {% do attrs.update({"name": name, "value": value}) %}
     <div x-data="dateTime('{{ event_name }}', '{{ value }}', {{ 'true' if time else 'false' }})"
-         class="w-full input">
+         class="w-full input" :class="{'has-focus': active}">
         <input class="w-full rounded-md border border-gray-200 px-3 py-2.5" x-ref="picker" type="text" {{ attributes(attrs) }}/>
     </div>
 {% endmacro %}
