+++ {{cookiecutter.repo_name}}/{{cookiecutter.repo_name}}/templates/forms/password.jinja
@@ -3,7 +3,7 @@
     <div x-data="password('update-{{ event_name | replace('_', '-') }}', '{{ value }}', 'password')" x-ref="input"
         :class="active ? 'has-focus' : ''"
          class="w-full z-10 flex items-center relative password">
-        <input  @focus="updateFlag" @blur="updateFlag" x-model.debounce.500ms="value" :type="type" name="{{ name }}" class="mr-8" {% if value != None %}value="{{ value }}"{% endif %} {{ attributes(attrs)}} />
+        <input x-model.debounce.500ms="value" :type="type" name="{{ name }}" class="mr-8" {% if value != None %}value="{{ value }}"{% endif %} {{ attributes(attrs)}} />
         <div class="absolute inset-y-0 right-4 flex items-center cursor-pointer"
              @click="type='password'"
              x-cloak
