+++ {{cookiecutter.repo_name}}/{{cookiecutter.repo_name}}/templates/forms/textarea.jinja
@@ -5,7 +5,6 @@
         x-ref="input"
          class="w-full input">
         <textarea
-            @focus="updateFlag" @blur="updateFlag"
             x-model.debounce.500ms="value"
             {% do attrs.update({"type": type, "name": name, "value": value}) %}
             {{ attributes(attrs)}}
